+++ packages/markdown-editor/src/index.ts
@@ -2,5 +2,6 @@ export { serializeToMdx } from './lib/serialize-to-mdx';
 export type {
   EditorBlock,
   EditorJsListItem,
+  SerializeOptions,
   SerializeToMdxInput,
 } from './types/serializer';